Fuzzing the number of subscribers to a sub.

+++ src/Subs/Subs/Sub.cs
@@ -23,7 +23,7 @@ namespace Subs
 
         public string Description { get; set; }
 
-        public ulong NumberOfSubscribers { get; set; }
+        public int NumberOfSubscribers { get; set; }
 
         public int Type { get; set; }
